+++ components/EntryTypeIndicator/index.tsx
@@ -3,6 +3,7 @@ import IconMinus from 'icons/minus.tsx';
 import IconCircle from 'icons/circle.tsx';
 import IconSquare from 'icons/square.tsx';
 import IconCheckbox from 'icons/checkbox.tsx';
+import { ICON_STANDARD } from 'styles';
 
 interface iEntryTypeIndicator {
   tags?: string[];
@@ -12,16 +13,16 @@ export function EntryTypeIndicator(props: iEntryTypeIndicator) {
   const { tags } = props;
 
   if (!tags) {
-    return <IconMinus />;
+    return <IconMinus class={ICON_STANDARD} />;
   } else if (tags.includes('permanent')) {
-    return <IconAsteriskSimple />;
+    return <IconAsteriskSimple class={ICON_STANDARD} />;
   } else if (tags.includes('event')) {
-    return <IconCircle />;
+    return <IconCircle class={ICON_STANDARD} />;
   } else if (tags.includes('task') && !tags.includes('done')) {
-    return <IconSquare />;
+    return <IconSquare class={ICON_STANDARD} />;
   } else if (tags.includes('task') && tags.includes('done')) {
-    return <IconCheckbox />;
+    return <IconCheckbox class={ICON_STANDARD} />;
   } else {
-    return <IconMinus />;
+    return <IconMinus class={ICON_STANDARD} />;
   }
 }